Initial, unrefined, version of ButtonBox.
This version is standalone, with no communication to a host.  Instead
alternately pulses one button LED and one indicator LED.  Pressing a
button changes the pulsing pair to the pressed button and its adjacent
indicator.

Dimming is also performed, using a SPDT momentary switch.  When
dimming is commanded, all LEDs flash momentarily to indicate the new
brightness.  This is also a nice test to be sure all LEDs are working.

All controls are debounced.

The LEDs are faded on and off during pulsing, a little like
incandescent bulbs, but not for the dimming preview.  This make the
viewing experience a bit less jarring than sudden step function
brightness changes.