Modify ButtonBox to interact with ButtonBox host.
ButtonBox sends and receives simple one byte message.  When a button
is pressed, the pinNumber is sent.  When a button is released, the
pinNumber is sent, with the high bit set.

When ButtonBox receives a byte, if the high bit is not set, then it
turns on the LED for the pinNumber in the byte.  If the high bit is
set, it turns of the LED.

If no message is received in 500 ms, the status indicator switches
from green to red, indicating loss-of-signal.